Utility for array operations

+++ main.cpp
@@ -4,14 +4,14 @@
 
 #include "include/data_table.hpp"
 #include "include/md_static/md_static_array.hpp"
-#include "include/md_static/md_static_mat_utility.hpp"
 #include "include/md_static/md_static_utility.hpp"
 
 int main(int argc, const char **argv) {
     size_t dm_size = 32;
-    MdStaticArray<int> f(
+    MdStaticArray<double> f(
         std::vector<size_t>({dm_size, dm_size, dm_size, dm_size}), 1);
-    MdStaticArray<int> g(std::vector<size_t>({dm_size, dm_size, dm_size}), 1);
+    MdStaticArray<double> g(std::vector<size_t>({dm_size, dm_size, dm_size}),
+                            1);
 
     // for (int i = 0; i < f.get_shape()[0]; ++i) {
     //     for (int j = 0; j < f.get_shape()[1]; ++j) {
@@ -26,7 +26,7 @@ int main(int argc, const char **argv) {
     //     }
     // }
     auto start = std::chrono::system_clock::now();
-    auto c = MdUtility::dot<int, int, int>(f, g);
+    auto c = MdUtility::dot<double, double, double>(f, g);
     // auto c = f + g;
     auto end = std::chrono::system_clock::now();
